added personUI.  didn't get as far as I wanted to unfortuntly

+++ Kid.java
@@ -9,7 +9,7 @@ public class Kid extends Person{
     /**
      * initializes the kid with name, age, phonenumber and favoriteCandy
      */
-    public Kid(String name, int age, int phoneNumber, String favoriteCandy){
+    public Kid(String name, int age, long phoneNumber, String favoriteCandy){
         super(name, age, phoneNumber);
         this.favoriteCandy = favoriteCandy;
     }
